hold problems
keycode in question: TD(TD_LY)
everything compiles fine.

tap (slash) and double tap (switch to qwerty and back) functions work well.

hold is supposed to turn on layer 2 momentarily, but instead, it just stays on, which i temporarily worked around by assigning TO(0) on layer 2 in that spot.

tap and then hold is supposed to turn layer 3 (marked as steno) on/off but just... doesn't do anything lmao

obviously the """steno""" is wip lol just trying to get it to work rn (i test it vs qwerty by tapping the spacebars as they're not supposed to output if i'm on the steno layer)